+++ lib/caching/memoize.ts
@@ -1,5 +1,3 @@
-/* eslint-disable @typescript-eslint/ban-types */
-
 /**
  * Turn a function into a memoized function. An optional resolver function can be passed which allows custom cache key generation.
  *
@@ -9,17 +7,22 @@
  * @param fn - Function to memoize
  * @param resolver - Optional resolver function to generate cache key. If not passed the first argument is used as map key
  */
-function memoize (fn:Function, resolver?:Function):Function {
-    const memoized = function () {
-        const key = typeof resolver === 'function' ? resolver.apply(this, arguments) : arguments[0]; // eslint-disable-line
+function memoize <
+    T extends (...args:any[]) => unknown,
+> (fn:T, resolver?:(...args:Parameters<T>) => any):T {
+    const memoized = function (...args:Parameters<T>) {
+        const key = typeof resolver === 'function' ? resolver(...args) : args[0];
         if (memoized.cache.has(key)) return memoized.cache.get(key);
 
-        const result = fn.apply(this, arguments); // eslint-disable-line
+        const result = fn(...args);
         memoized.cache.set(key, result);
         return result;
     };
     memoized.cache = new Map();
-    return memoized;
+
+    /* eslint-disable-next-line */
+    /* @ts-ignore */
+    return memoized as T;
 }
 
 export {memoize, memoize as default};